충남대 BE_박지인 - 1단계 테스트 코드 작성 (#168)
* refactor: update 메서드 리팩토링

* feat(validation): 검증을 위한 annotation 생성

* chore: validation 의존성 추가

* feat: validation 적용 및 Request DTO 분리

* feat: 검증 적용 및 update의 request DTO 변경

* refactor: DTO 분리에 따른 update 메서드 파라미터 수정

* feat: 전역 예외 처리 구현

* docs(README): README 작성

* refactor: 코드 수정

* test: 테스트 코드 작성

---------

+++ src/main/java/gift/domain/Product.java
@@ -4,16 +4,16 @@ public class Product {
 
     Long id;
     String name;
-    Integer price;
+    Long price;
     String imageURL;
 
-    public Product(String name, Integer price, String imageURL) {
+    public Product(String name, Long price, String imageURL) {
         this.name = name;
         this.price = price;
         this.imageURL = imageURL;
     }
 
-    public Product(Long id, String name, Integer price, String imageURL) {
+    public Product(Long id, String name, Long price, String imageURL) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -28,7 +28,7 @@ public class Product {
         return name;
     }
 
-    public Integer getPrice() {
+    public Long getPrice() {
         return price;
     }
