+++ src/main/java/gift/dto/request/ProductUpdateReqDTO.java
@@ -10,7 +10,7 @@ public record ProductUpdateReqDTO(
     @PositiveOrZero(
         message = "올바른 상품 가격을 입력해주세요."
     )
-    Integer price,
+    Long price,
 
     String imageURL
 ) {
