+++ src/main/java/gift/dto/response/ProductResDTO.java
@@ -1,5 +1,5 @@
 package gift.dto.response;
 
-public record ProductResDTO(Long id, String name, int price, String imageURL) {
+public record ProductResDTO(Long id, String name, Long price, String imageURL) {
 
 }
